필터 컴포넌트로 수정
